requirements.txt
openai
faiss-cpu
fastapi
uvicorn
PyPDF2
python-dotenv
